Safari-609.1.11.1/LayoutTests/http/tests/misc/redirect-to-external-url-expected.txt

Policy delegate: attempt to load http://127.0.0.1:8000/misc/resources/redirect-to-external-url.php with navigation type 'link clicked' originating from HTML > #document
Policy delegate: attempt to load spaceballs://the-flamethrower/ with navigation type 'link clicked' originating from HTML > #document
Policy delegate: unable to implement policy with error domain 'WebKitErrorDomain', error code 101, in frame '<!--frame1-->'
This page has an iframe. It loads then navigates to an http url which redirects to an externally handled url.
The navigation policy delegate should be consulted about the http url and the external url.


--------
Frame: '<!--frame1-->'
--------
This is a page with a link in it. When the link it clicked it will navigate to another page. That page redirects to an externally handled URL.
The policy delegate should be asked about the initial navigation and the redirect.
The kids love this one
